@@ -10,28 +10,30 @@
         "cartalyst/sentry": "2.1.*",
         "doctrine/dbal": "~2.3",
         "ezyang/htmlpurifier": "dev-master",
+        "guzzlehttp/guzzle": "~6.2",
         "illuminate/container": "4.0.*@dev",
         "illuminate/database": "4.0.*@dev",
         "illuminate/events": "4.0.*@dev",
         "illuminate/support": "4.0.*@dev",
         "intervention/image": "1.5.*",
         "ircmaxell/random-lib": "~1.2",
-        "kzykhys/ciconia": "~1.0",
+        "michelf/php-markdown": "~1.7",
         "monolog/monolog": "~1.22",
         "pagerfanta/pagerfanta": "1.0.*@dev",
         "pimple/pimple": "3.*",
         "robmorgan/phinx": "*",
         "silex/silex": "2.*",
         "swiftmailer/swiftmailer": "5.0",
-        "symfony/config": "~2.3",
-        "symfony/console": "~2.3",
-        "symfony/css-selector": "~2.3",
-        "symfony/form": "~2.3",
-        "symfony/intl": "~2.3",
-        "symfony/security-csrf": "~2.3",
-        "symfony/translation": "~2.3",
-        "symfony/twig-bridge": "~2.3",
-        "symfony/validator": "~2.3",
+        "symfony/config": "~3.2",
+        "symfony/console": "~3.2",
+        "symfony/css-selector": "~3.2",
+        "symfony/form": "~3.2",
+        "symfony/intl": "~3.2",
+        "symfony/security": "~3.2",
+        "symfony/security-csrf": "~3.2",
+        "symfony/translation": "~3.2",
+        "symfony/twig-bridge": "~3.2",
+        "symfony/validator": "~3.2",
         "symfony/yaml": "~2.5",
         "twig/twig": "1.*",
         "vlucas/spot2": "dev-master"
@@ -40,10 +42,10 @@
         "codeclimate/php-test-reporter": "0.2.0",
         "friendsofphp/php-cs-fixer": "^2.0.0",
         "fzaninotto/faker": "^1.5.0",
-        "johnkary/phpunit-speedtrap": "~1.0@dev",
+        "johnkary/phpunit-speedtrap": "2.0.x-dev",
         "mockery/mockery": "1.0.*@dev",
-        "phpunit/dbunit": "1.3.2",
-        "phpunit/phpunit": "^4.8",
+        "phpunit/phpunit": "~6.0",
+        "phpunit/phpunit-mock-objects": "~4.0",
         "satooshi/php-coveralls": "~0.6"
     },
     "autoload": {
