@@ -5,7 +5,7 @@ namespace OpenCFP\Test;
 use Mockery as m;
 use OpenCFP\Application;
 
-class ContainerAwareTest extends \PHPUnit_Framework_TestCase
+class ContainerAwareTest extends \PHPUnit\Framework\TestCase
 {
     public function testAllowsToRetrieveService()
     {
