@@ -4,7 +4,7 @@ namespace OpenCFP\Test\Domain\Talk;
 
 use OpenCFP\Domain\Talk\TalkSubmission;
 
-class TalkSubmissionTest extends \PHPUnit_Framework_TestCase
+class TalkSubmissionTest extends \PHPUnit\Framework\TestCase
 {
     /** @test */
     public function it_should_be_created_from_native_format()
@@ -36,10 +36,11 @@ class TalkSubmissionTest extends \PHPUnit_Framework_TestCase
     /**
      * @test
      * @dataProvider invalidTalkTitles
+     * @expectedException \OpenCFP\Domain\Talk\InvalidTalkSubmissionException
+     * @expectedExceptionMessage title
      */
     public function it_guards_that_title_is_appropriate_length($title)
     {
-        $this->setExpectedException(\OpenCFP\Domain\Talk\InvalidTalkSubmissionException::class, 'title');
         TalkSubmission::fromNative(['title' => $title]);
     }
 
@@ -51,20 +52,26 @@ class TalkSubmissionTest extends \PHPUnit_Framework_TestCase
         ];
     }
 
-    /** @test */
+    /**
+     * @test
+     * @expectedException \OpenCFP\Domain\Talk\InvalidTalkSubmissionException
+     * @expectedExceptionMessage description
+     */
     public function it_guards_that_description_is_provided()
     {
-        $this->setExpectedException(\OpenCFP\Domain\Talk\InvalidTalkSubmissionException::class, 'description');
         TalkSubmission::fromNative([
             'title' => 'Talk With No Description',
             'description' => '',
         ]);
     }
 
-    /** @test */
+    /**
+     * @test
+     * @expectedException \OpenCFP\Domain\Talk\InvalidTalkSubmissionException
+     * @expectedExceptionMessage talk type
+     */
     public function it_guards_that_invalid_talk_types_cannot_be_used()
     {
-        $this->setExpectedException(\OpenCFP\Domain\Talk\InvalidTalkSubmissionException::class, 'talk type');
         TalkSubmission::fromNative([
             'title' => 'Some off-the-wall Talk Type',
             'description' => 'I do not play by the rules.',
@@ -72,10 +79,13 @@ class TalkSubmissionTest extends \PHPUnit_Framework_TestCase
         ]);
     }
 
-    /** @test */
+    /**
+     * @test
+     * @expectedException \OpenCFP\Domain\Talk\InvalidTalkSubmissionException
+     * @expectedExceptionMessage level
+     */
     public function it_guards_that_invalid_level_cannot_be_used()
     {
-        $this->setExpectedException(\OpenCFP\Domain\Talk\InvalidTalkSubmissionException::class, 'level');
         TalkSubmission::fromNative([
             'title' => 'Invalid Skill Level Talk',
             'description' => 'I do not play by the rules.',
@@ -84,10 +94,13 @@ class TalkSubmissionTest extends \PHPUnit_Framework_TestCase
         ]);
     }
 
-    /** @test */
+    /**
+     * @test
+     * @expectedException \OpenCFP\Domain\Talk\InvalidTalkSubmissionException
+     * @expectedExceptionMessage category
+     */
     public function it_guards_that_invalid_categories_cannot_be_assigned()
     {
-        $this->setExpectedException(\OpenCFP\Domain\Talk\InvalidTalkSubmissionException::class, 'category');
         TalkSubmission::fromNative([
             'title' => 'Invalid Categorized Talk',
             'description' => 'I do not play by the rules.',
