@@ -7,7 +7,7 @@ use OpenCFP\Environment;
 /**
  * @covers OpenCFP\Environment
  */
-class EnvironmentTest extends \PHPUnit_Framework_TestCase
+class EnvironmentTest extends \PHPUnit\Framework\TestCase
 {
     /** @test */
     public function it_should_encapsulate_valid_environments()
@@ -32,11 +32,12 @@ class EnvironmentTest extends \PHPUnit_Framework_TestCase
         $this->assertEquals('testing', Environment::fromString('testing'));
     }
 
-    /** @test */
+    /**
+     * @test
+     * @expectedException \InvalidArgumentException
+     */
     public function it_fails_when_given_an_invalid_environment_string()
     {
-        $this->setExpectedException('InvalidArgumentException');
-
         Environment::fromString('foo');
     }
 }
