@@ -10,7 +10,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Session\Session;
 use Symfony\Component\HttpFoundation\Session\Storage\MockFileSessionStorage;
 
-class SpeakersControllerTest extends \PHPUnit_Framework_TestCase
+class SpeakersControllerTest extends \PHPUnit\Framework\TestCase
 {
     private $app;
 
