@@ -8,7 +8,7 @@ use OpenCFP\Domain\Talk\TalkRepository;
 use OpenCFP\Infrastructure\Persistence\SpotTalkRepository;
 use Spot\Mapper;
 
-class SpotTalkRepositoryTest extends \PHPUnit_Framework_TestCase
+class SpotTalkRepositoryTest extends \PHPUnit\Framework\TestCase
 {
     public function testImplementsTalkRepository()
     {
@@ -17,23 +17,6 @@ class SpotTalkRepositoryTest extends \PHPUnit_Framework_TestCase
         $repository = new SpotTalkRepository($mapper);
 
         $this->assertInstanceOf(TalkRepository::class, $repository);
-    }
-
-    public function testPersistSavesTalk()
-    {
-        $talk = $this->getTalkMock();
-
-        $mapper = $this->getMapperMock();
-
-        $mapper
-            ->shouldReceive('save')
-            ->once()
-            ->with($talk)
-        ;
-
-        $repository = new SpotTalkRepository($mapper);
-
-        $repository->persist($talk);
     }
 
     //
@@ -46,13 +29,5 @@ class SpotTalkRepositoryTest extends \PHPUnit_Framework_TestCase
     private function getMapperMock()
     {
         return m::mock(Mapper::class);
-    }
-
-    /**
-     * @return m\MockInterface|Entity\Talk
-     */
-    private function getTalkMock()
-    {
-        return m::mock(Entity\Talk::class);
     }
 }
