@@ -5,7 +5,7 @@ namespace OpenCFP\Test\Provider;
 use OpenCFP\Provider\SymfonySentrySession;
 use Symfony\Component\HttpFoundation\Session\SessionInterface;
 
-class SymfonySentrySessionTest extends \PHPUnit_Framework_TestCase
+class SymfonySentrySessionTest extends \PHPUnit\Framework\TestCase
 {
     public function testDefaults()
     {
