@@ -4,7 +4,7 @@ namespace OpenCFP\Test\Util\Faker;
 
 use Faker\Generator;
 
-class GeneratorTest extends \PHPUnit_Framework_TestCase
+class GeneratorTest extends \PHPUnit\Framework\TestCase
 {
     use GeneratorTrait;
 
