@@ -4,17 +4,17 @@ namespace OpenCFP\Test\Util;
 
 use Exception;
 
-class ImmutableTest extends \PHPUnit_Framework_TestCase
+class ImmutableTest extends \PHPUnit\Framework\TestCase
 {
-
-    /** @test */
+    /**
+     * @test
+     * @expectedException Exception
+     * @expectedExceptionMessage ImmutableObject is immutable
+     */
     public function it_throws_exception_when_assignment_is_made_after_construction()
     {
         $object = new ImmutableObject('Foo');
-
         $this->assertEquals('Foo', $object->value);
-
-        $this->setExpectedException(Exception::class, 'ImmutableObject is immutable.');
         $object->value = 'Bar';
     }
 }
