@@ -3,22 +3,14 @@ namespace OpenCFP\Http\Form;
 
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\Validator\Constraints as Assert;
+use Symfony\Component\Form\Extension\Core\Type\EmailType;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 
 class ForgotForm extends AbstractType
 {
     public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $builder->add('email', 'text', [
-            'constraints' => [
-                new Assert\NotBlank(),
-                new Assert\Email(),
-            ],
-        ]);
-    }
-
-    public function getName()
-    {
-        return 'forgot';
+        $builder->add('email', EmailType::class)
+            ->add('send', SubmitType::class, ['label' => 'Reset my password']);
     }
 }
